Add classes to Resistances area
Added classes to the containing div for the Resistances area, so we can tell when it's empty, and what type of content it has when full. (Used primarily for hiding the area when empty.)

ATM this outputs the class once for every instance it finds in the resistances array, which is less than ideal, but does work. Maybe at some stage I'll come up with an elegant way to avoid it, but the only ways I can think of right now would massively overcomplicate the template code, so I don't wanna.