Dev (#18)
* NPC sheet/chat card improvements

* Resistances area on the NPC sheet no longer takes up space while empty. This should make simple monsters a bit more compact.

* Hide NPC flavour on autocards

Since my "don't output flavour for auto-generated NPC chat cards" logic was reverted in the system module, I've replaced it with a simple bit of logic to collapse the flavour area on NPC chat cards by default.

* Add classes to Resistances area

Added classes to the containing div for the Resistances area, so we can tell when it's empty, and what type of content it has when full. (Used primarily for hiding the area when empty.)

ATM this outputs the class once for every instance it finds in the resistances array, which is less than ideal, but does work. Maybe at some stage I'll come up with an elegant way to avoid it, but the only ways I can think of right now would massively overcomplicate the template code, so I don't wanna.

* Added new script file

Added the chat.js scripts file to handle auto-collapse of NPC chat card flavour.