Updated usage display
Since the default for limited uses on new powers is now 0/0 instead of null/null, it was necessary to update how the logic handled output of different usage values. Which was deeply annoying to do with Handlebars, so I got the shits with it and just wrote a custom helper instead.